cheddar: Make a GCP service account and key
I use GCP quite a lot. In particular each machine gets its own service
account, and a private key for that account has to be readable somehow.

Provisioning secrets on NixOS is a pain in the ass[1] that continuously
manufactures Discourse threads, blog posts, and one-off tools. I think
the Right Thing is probably just minting the key as part of a machine
image, same as you would do with a SSH key/cert. Since I don't make
images automatically yet, I chose this instead: use sops to encrypt the
secret material - readable by the machine's SSH key - and leave the
encrypted version in source control (and the Nix store).

1. https://nixos.wiki/wiki/Comparison_of_secret_managing_schemes