flake: Set up Go development
This adds the Go tools to the development shell and a cute overlay[1]
that generates Nix hashes for Go module projects.

There is a minor mystery in default behavior here: the Go package in the
nixpkgs input is 1.19, as is buildPackages.go, but `nix build .#somepackage`
builds with Go 1.20. Setting `go = pkgs.go` does the right thing (as long as
it's compatible with what go.mod says).

1. https://www.tweag.io/blog/2021-03-04-gomod2nix/