hedwig: Use UniFi controller version 7.2.97
What a mess. Getting devices adopted to the new controller was hard
enough[1], but the controller itself was highly unstable and was prone to
Java NPEs and failed startup. The 6.x version recommends upgrading to
exactly 7.2.97, and that worked. I pinned the JRE to Java 8 because that's
what the release notes[2] say, and I'm now in the business of reading their
release notes very carefully every time.

The newest controller from nixpkgs-unstable was definitely working after
commit 4d57f13, so I'm not sure where all the spooky startup problems came
from. But this weekend was so screwed up that I'm going to be superstitious
about these updates forever now.

1. I had to set-inform manually on the devices for some reason; the
   documentation says that shouldn't be required after a factory reset, but
   I have hours of toil that say otherwise.
2. https://community.ui.com/releases/UniFi-Network-Application-7-2-97/d2f71152-3001-42cd-834c-9245d86f4e72