hedwig: Enable UniFi poller for Prometheus
I added my own unpoller module that mostly copies the one in nixpkgs, except
it makes the UniFi password option type str instead of path. Using a path
would make the password be readable in the Nix store. Once I'm sure this
works, I'll try to upstream a fix and get rid of it.