moneydance: Package for Nix
This minimally works but in an unsatisfying way.

Moneydance ships with a copy of OpenJDK 21. I tried to use that by doing the
patchelf thing but failed. Running `jre/bin/java -version` works, but
running Moneydance just writes a little output to the console and exits.
I'm probably missing something obvious.

The script shipped with the Moneydance installer pulls in some JavaFX
modules:

  --add-modules javafx.swing,javafx.controls,javafx.graphics

That doesn't work with OpenJDK 17 from Nix, because that doesn't include
OpenJFX. The Oracle JDK ships with JavaFX stuff, but Nix hasn't packaged new
versions of that.

Running Moneydance without the JavaFX modules seems to work, but who knows
if that will cause problems down the road. Based on just the names, I'm
guessing this is going to break charts and graphs.