prometheus: fire an alert if any of our SMART metrics is nonzero
This is a bit crude. Per https://www.backblaze.com/blog/what-smart-stats-indicate-hard-drive-failures/
we probably want to do some sort of rate(...[1d]) computation on most of these
metrics (other than 197) and alert off that, rather than just alerting on
anything non-zero. In fact, one of my drives (in a mirror, so low risk) as a
nonzero value for 187 (reported uncorrectable errors, it reports 10) and that
could be something as benign as a bad moment with cosmic rays.