trackphotos: Give up on NextCloud
That was horrific.

My use case should be simple: use the Photos app with an S3-style cloud
object store. Reader, this was not simple. The Photos app stores some things
in the object store and some other things on the local filesystem. The app
doesn't create its own root directory (!) automatically; the only way to
find out why photos are silently failing to show up after upload is to sniff
a nearly inscrutable PHP error message that is logged at debug (!!) level.
Once that directory is created, everything mostly seems to work, but the UI
is so slow and prone to random errors that I can never trust this with
actual content.

Also, the NixOS module tries mightily, but NextCloud resists having a
reproducible config at every turn. A website allegedly meant to run on a
server but scribbling on its own config file(s) is a thing of a morally and
intellectually backward time that I thought was behind us. The rest of us
have learned from our 1990s mistakes and moved on.