idiotbox: Set container service dependencies
This change should make the VPN container service depend on the service that
creates the Podman pod and make the idiotbox containers depend on both of
those. That is, the VPN service will wait for the pod service before trying
to start, and the idiotbox container services will be brought down if the
VPN service is down.