securenets: A module to set up a WireGuard mesh
This lets you specify a network as a list of hosts with optional information
about what site they live in and what static addresses they have. Then that
automatically generates a systemd-networkd[1] config on each host. That
config infers its own address from the network spec and creates peer config
for each other host in the spec.

This is still a bit messy, since you have to provision private keys out of
band, but one yak at a time.

1. https://www.freedesktop.org/software/systemd/man/latest/systemd-networkd.service.html