mail: Set all MX records to smtp.bergmans.us
A handful of connections to smtpd (maybe 10-20 per day) were failing with
the error "bad certificate...SSL alert number 42." I'm speculating that
that's because the content of domains' MX records didn't match the subject
name in the server certificate (smtp.bergmans.us). In the day that this has
been deployed, the error hasn't recurred. Closes #5.