bug(hedwig): Record the low 16 bits of SMART attr 188
Both Seagate drives in my array began alerting for nonzero values for
SMART attribute 188, with a gigantic raw value like 21475164165.
Shoutout to [1] for explaining that this is 0x000500050005, and the
total number of timeouts is in the low 16 bits. So this is 5 total
command timeouts, not billions.

Unraid documentation says[2] that for attribute 188, "Occasional
timeouts can happen due to power problems. If this is frequently above
0, check your cables and power connections." I just replaced the PSU on
this machine, so that tracks.

The WD drives don't report 188 at all, but if some other manufacturer
does, then we'll have to special-case the handling of this value by
drive model like smartmontools does[3].

1. https://superuser.com/a/1747851
2. https://docs.unraid.net/unraid-os/system-administration/monitor-performance/smart-reports-and-disk-health/
3. https://github.com/smartmontools/smartmontools/blob/main/lib/drivedb.h